refactor pollbook package configuration into seperate files
