+++ backend/src/app.ts
@@ -1,42 +1,24 @@
 import * as grout from '@votingworks/grout';
 import express, { Application } from 'express';
 import { err, ok, Result } from '@votingworks/basics';
-import { UsbDrive } from '@votingworks/usb-drive';
-import { readFile, ReadFileError } from '@votingworks/fs';
-import { join } from 'node:path';
-import {
-  getEntries,
-  getFileByName,
-  isElectionManagerAuth,
-  openZip,
-  readTextEntry,
-} from '@votingworks/utils';
-import {
-  DEFAULT_SYSTEM_SETTINGS,
-  PrinterStatus,
-  safeParseJson,
-} from '@votingworks/types';
-import { parse } from 'csv-parse/sync';
+import { DEFAULT_SYSTEM_SETTINGS, PrinterStatus } from '@votingworks/types';
 import { setInterval } from 'node:timers/promises';
-import {
-  DippedSmartCardAuthApi,
-  DippedSmartCardAuthMachineState,
-} from '@votingworks/auth';
-import { Printer, renderToPdf } from '@votingworks/printing';
+import { DippedSmartCardAuthMachineState } from '@votingworks/auth';
+import { renderToPdf } from '@votingworks/printing';
 import React from 'react';
 import { getBatteryInfo } from '@votingworks/backend';
-import { Workspace } from './workspace';
 import {
+  Workspace,
+  AppContext,
   DeviceStatuses,
   Election,
-  ElectionSchema,
   MachineInformation,
   PollbookConnectionStatus,
   PollbookEvent,
-  PollbookPackage,
   Voter,
   VoterIdentificationMethod,
   VoterSearchParams,
+  ConfigurationStatus,
 } from './types';
 import { AvahiService } from './avahi';
 import { rootDebug } from './debug';
@@ -46,31 +28,9 @@ import {
   PORT,
 } from './globals';
 import { CheckInReceipt } from './check_in_receipt';
-import { HlcTimestamp, HybridLogicalClock } from './hybrid_logical_clock';
+import { pollUsbDriveForPollbookPackage } from './pollbook_package';
 
 const debug = rootDebug;
-const usbDebug = debug.extend('usb');
-
-export interface AppContext {
-  workspace: Workspace;
-  auth: DippedSmartCardAuthApi;
-  usbDrive: UsbDrive;
-  printer: Printer;
-  machineId: string;
-}
-
-const MEGABYTE = 1024 * 1024;
-const MAX_POLLBOOK_PACKAGE_SIZE = 10 * MEGABYTE;
-
-function toCamelCase(str: string) {
-  const words = str
-    .split(/[^a-zA-Z0-9]/)
-    .filter((word) => word.length > 0)
-    .map((word) => word.toLowerCase());
-  const first = words.shift();
-  const rest = words.map((word) => word[0].toUpperCase() + word.slice(1));
-  return [first, ...rest].join('');
-}
 
 function createApiClientForAddress(address: string): grout.Client<Api> {
   debug('Creating API client for address %s', address);
@@ -91,101 +51,6 @@ function constructAuthMachineState(
       date: election.date,
     },
   };
-}
-
-async function readPollbookPackage(
-  path: string
-): Promise<Result<PollbookPackage, ReadFileError>> {
-  const pollbookPackage = await readFile(path, {
-    maxSize: MAX_POLLBOOK_PACKAGE_SIZE,
-  });
-  if (pollbookPackage.isErr()) {
-    return err(pollbookPackage.err());
-  }
-  const zipFile = await openZip(pollbookPackage.ok());
-  const zipName = 'pollbook package';
-  const entries = getEntries(zipFile);
-
-  const electionEntry = getFileByName(entries, 'election.json', zipName);
-  const electionJsonString = await readTextEntry(electionEntry);
-  const election: Election = safeParseJson(
-    electionJsonString,
-    ElectionSchema
-  ).unsafeUnwrap();
-
-  const votersEntry = getFileByName(entries, 'voters.csv', zipName);
-  const votersCsvString = await readTextEntry(votersEntry);
-  const voters = parse(votersCsvString, {
-    columns: (header) => header.map(toCamelCase),
-    skipEmptyLines: true,
-    // Filter out metadata row at the end
-    onRecord: (record) => (record.voterId ? record : null),
-  }) as Voter[];
-
-  return ok({ election, voters });
-}
-
-type ConfigurationStatus = 'loading' | 'not-found';
-let configurationStatus: ConfigurationStatus | undefined;
-
-function pollUsbDriveForPollbookPackage({
-  auth,
-  workspace,
-  usbDrive,
-}: AppContext) {
-  usbDebug('Polling USB drive for pollbook package');
-  if (workspace.store.getElection()) {
-    return;
-  }
-  process.nextTick(async () => {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    for await (const _ of setInterval(100)) {
-      const usbDriveStatus = await usbDrive.status();
-      if (usbDriveStatus.status !== 'mounted') {
-        continue;
-      }
-      usbDebug('Found USB drive mounted at %s', usbDriveStatus.mountPoint);
-
-      const authStatus = await auth.getAuthStatus(
-        constructAuthMachineState(workspace)
-      );
-      if (!isElectionManagerAuth(authStatus)) {
-        usbDebug('Not logged in as election manager, not configuring');
-        continue;
-      }
-
-      configurationStatus = 'loading';
-      const pollbookPackageResult = await readPollbookPackage(
-        join(usbDriveStatus.mountPoint, 'pollbook-package.zip')
-      );
-      if (pollbookPackageResult.isErr()) {
-        const result = pollbookPackageResult.err();
-        debug('Read pollbook package error: %O', result);
-        if (
-          result.type === 'OpenFileError' &&
-          'code' in result.error &&
-          result.error.code === 'ENOENT'
-        ) {
-          configurationStatus = 'not-found';
-        } else {
-          throw result;
-        }
-        configurationStatus = 'not-found';
-        continue;
-      }
-      const pollbookPackage = pollbookPackageResult.ok();
-      workspace.store.setElectionAndVoters(
-        pollbookPackage.election,
-        pollbookPackage.voters
-      );
-      configurationStatus = undefined;
-      debug('Configured with pollbook package: %O', {
-        election: pollbookPackage.election,
-        voters: pollbookPackage.voters.length,
-      });
-      break;
-    }
-  });
 }
 
 async function setupMachineNetworking({
@@ -366,6 +231,7 @@ function buildApi(context: AppContext) {
     },
 
     getElection(): Result<Election, 'unconfigured' | ConfigurationStatus> {
+      const configurationStatus = store.getConfigurationStatus();
       if (configurationStatus) {
         return err(configurationStatus);
       }
