+++ backend/src/auth.ts
@@ -1,0 +1,16 @@
+import { DippedSmartCardAuthMachineState } from '@votingworks/auth';
+import { DEFAULT_SYSTEM_SETTINGS } from '@votingworks/types';
+import { Workspace } from './types';
+
+export function constructAuthMachineState(
+  workspace: Workspace
+): DippedSmartCardAuthMachineState {
+  const election = workspace.store.getElection();
+  return {
+    ...DEFAULT_SYSTEM_SETTINGS['auth'],
+    electionKey: election && {
+      id: election.id,
+      date: election.date,
+    },
+  };
+}
