+++ backend/src/globals.ts
@@ -36,3 +36,6 @@ export const MACHINE_DISCONNECTED_TIMEOUT = 60000;
 export const NETWORK_REQUEST_TIMEOUT = 1000;
 
 export const NETWORK_EVENT_LIMIT = 500;
+
+export const MEGABYTE = 1024 * 1024;
+export const MAX_POLLBOOK_PACKAGE_SIZE = 10 * MEGABYTE;
