+++ backend/src/server.ts
@@ -1,8 +1,9 @@
 import { useDevDockRouter } from '@votingworks/dev-dock-backend';
 import express from 'express';
 import { CITIZEN_THERMAL_PRINTER_CONFIG } from '@votingworks/printing';
-import { AppContext, buildApp } from './app';
+import { buildApp } from './app';
 import { PORT } from './globals';
+import { AppContext } from './types';
 
 /**
  * Starts the server.
