+++ backend/src/workspace.ts
@@ -2,12 +2,8 @@ import { ensureDirSync } from 'fs-extra';
 import { join } from 'node:path';
 
 import { BaseLogger } from '@votingworks/logging';
+import { Workspace } from './types';
 import { Store } from './store';
-
-export interface Workspace {
-  assetDirectoryPath: string;
-  store: Store;
-}
 
 export function createWorkspace(
   workspacePath: string,
